+++ src/main/java/net/silentchaos512/gear/init/ModLootStuff.java
@@ -18,17 +18,22 @@
 
 package net.silentchaos512.gear.init;
 
+import net.minecraft.world.storage.loot.conditions.LootConditionManager;
 import net.minecraft.world.storage.loot.functions.LootFunctionManager;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.LootTableLoadEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.silentchaos512.gear.loot.SelectGearTierLootFunction;
-import net.silentchaos512.gear.loot.SetPartsFunction;
+import net.silentchaos512.gear.loot.condition.HasPartCondition;
+import net.silentchaos512.gear.loot.condition.HasTraitCondition;
+import net.silentchaos512.gear.loot.function.SelectGearTierLootFunction;
+import net.silentchaos512.gear.loot.function.SetPartsFunction;
 
 public final class ModLootStuff {
     private ModLootStuff() {}
 
     public static void init() {
+        LootConditionManager.registerCondition(HasPartCondition.SERIALIZER);
+        LootConditionManager.registerCondition(HasTraitCondition.SERIALIZER);
         LootFunctionManager.registerFunction(SelectGearTierLootFunction.SERIALIZER);
         LootFunctionManager.registerFunction(SetPartsFunction.SERIALIZER);