+++ src/main/java/net/silentchaos512/gear/loot/function/SelectGearTierLootFunction.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.silentchaos512.gear.loot;
+package net.silentchaos512.gear.loot.function;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonObject;